requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.107.0
google-auth-httplib2==0.2.0
pydantic==2.5.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.6
httpx==0.25.2
pydantic-core==2.14.1
requests==2.31.0
numpy==1.24.3
scipy==1.11.4
torch==2.1.1
scikit-learn==1.3.2
